ref: move to interconnect module
