+++ src/interconnect/azext_interconnect/aaz/latest/interconnect_group/_delete.py
