+++ src/interconnect/azext_interconnect/aaz/latest/interconnect_group/_get_node_availability.py
