+++ src/interconnect/azext_interconnect/tests/latest/recordings/test_interconnect_group_crud.yaml
