+++ src/interconnect/azext_interconnect/tests/latest/test_interconnect_group.py
