+++ src/service_name.json
@@ -1073,5 +1073,10 @@
     "Command": "az changesafety",
     "AzureServiceName": "ChangeSafety",
     "URL": ""
+  },
+  {
+    "Command": "az interconnect-group",
+    "AzureServiceName": "Interconnect",
+    "URL": ""
   }
 ]
